Fixed Bug
Bug when an enemy attacks sometimes the game would crash because of a index out of bound error.

While this solves the problem not sure if its the right solution.
The battleresult.crited arraylist is supposed to have an item added to it when the battle is calculated, but sometimes in some branch no item ever gets added to the array list. I added code in the initilization of the battle result at the very end to make sure if the crited array list is empty that a false item is added to make sure it doesnt crash later.